Inhibit AST rewriting inside 'async def' functions.
Remember, our goal is to make async code friendlier in interactive
ipython sessions without making it incompatible or semantically
different.

We want to improve top-level calls:

>>> do_an_async_thing()  # this should return a Task, not a coroutine,
...                      # since it was called at the top level of an
...                      # interactive session and a coroutine() object
...                      # is not ergonomic

However, for functions defined in an ipython cell it's less clear what
to do:

>>> def sync_call():
...     do_an_async_thing()      # this produces a coroutine() object
...                              # inside a synchronous function - we
...                              # could swap it out for a Task, but we
...                              # couldn't await it.

>>> async def async_call():
...     c = do_an_async_thing()  # this also produces a coroutine()
...                              # object before we do any AST
...                              # transformations. We could swap it for
...                              # a Task and await it, just like we do
...                              # for top-level coroutines.

In both of these cases, it's better to be less intrusive. We're inside
a function and can expect the author to be thinking "Python code" not
"interactive environment". Moreover, converting the coroutine into a
Task schedules it - which means code copied from an ipython cell to an
external Python script (a library, or whatever) would behave
differently.

Better to be less intrusive and more predictable.