Correct bogus stack-trace information after AST traversal (fixes #2)
We splice in AST data generated by ast.parse()ing a text block.  This
AST carries its own metadata (lineno and friends) associated with the
injected code. When it's introspected in the process of generating a
traceback, it's bogus because it doesn't correspond to lines of code in
the actual ipython cell.

This patch introduces a second AST transformer that recursively copies
node metadata from the original, un-transformed AST. As a result,
traceback can point to the correct line of original code (even if it's
not what actually got executed.)